requirements.txt
streamlit
pandas
google-generativeai
gspread
oauth2client